Updates the colors/typography theme files.
