--- boilerplate/app/theme/fonts.ts
@@ -1,44 +1,0 @@
-import {
-  SpaceGrotesk_300Light,
-  SpaceGrotesk_400Regular,
-  SpaceGrotesk_500Medium,
-  SpaceGrotesk_600SemiBold,
-  SpaceGrotesk_700Bold,
-} from "@expo-google-fonts/space-grotesk"
-
-export const customFontsToLoad = {
-  SpaceGrotesk_300Light,
-  SpaceGrotesk_400Regular,
-  SpaceGrotesk_500Medium,
-  SpaceGrotesk_600SemiBold,
-  SpaceGrotesk_700Bold,
-}
-
-export const fonts = {
-  // Cross-platform Google font.
-  spaceGrotesk: {
-    light: "SpaceGrotesk_300Light",
-    normal: "SpaceGrotesk_400Regular",
-    medium: "SpaceGrotesk_500Medium",
-    semiBold: "SpaceGrotesk_600SemiBold",
-    bold: "SpaceGrotesk_700Bold",
-  },
-  // iOS only font.
-  helveticaNeue: {
-    thin: "HelveticaNeue-Thin",
-    light: "HelveticaNeue-Light",
-    normal: "Helvetica Neue",
-    medium: "HelveticaNeue-Medium",
-  },
-  // iOS only font.
-  courier: { normal: "Courier" },
-  // Android only font.
-  sansSerif: {
-    thin: "sans-serif-thin",
-    light: "sans-serif-light",
-    normal: "sans-serif",
-    medium: "sans-serif-medium",
-  },
-  // Android only font.
-  monospace: { normal: "monospace" },
-}
