+++ boilerplate/app/theme/typography.ts
@@ -1,5 +1,52 @@
 import { Platform } from "react-native"
-import { fonts } from "./fonts"
+import {
+  SpaceGrotesk_300Light,
+  SpaceGrotesk_400Regular,
+  SpaceGrotesk_500Medium,
+  SpaceGrotesk_600SemiBold,
+  SpaceGrotesk_700Bold,
+} from "@expo-google-fonts/space-grotesk"
+
+export const customFontsToLoad = {
+  SpaceGrotesk_300Light,
+  SpaceGrotesk_400Regular,
+  SpaceGrotesk_500Medium,
+  SpaceGrotesk_600SemiBold,
+  SpaceGrotesk_700Bold,
+}
+
+const fonts = {
+  spaceGrotesk: {
+    // Cross-platform Google font.
+    light: "SpaceGrotesk_300Light",
+    normal: "SpaceGrotesk_400Regular",
+    medium: "SpaceGrotesk_500Medium",
+    semiBold: "SpaceGrotesk_600SemiBold",
+    bold: "SpaceGrotesk_700Bold",
+  },
+  helveticaNeue: {
+    // iOS only font.
+    thin: "HelveticaNeue-Thin",
+    light: "HelveticaNeue-Light",
+    normal: "Helvetica Neue",
+    medium: "HelveticaNeue-Medium",
+  },
+  courier: {
+    // iOS only font.
+    normal: "Courier",
+  },
+  sansSerif: {
+    // Android only font.
+    thin: "sans-serif-thin",
+    light: "sans-serif-light",
+    normal: "sans-serif",
+    medium: "sans-serif-medium",
+  },
+  monospace: {
+    // Android only font.
+    normal: "monospace",
+  },
+}
 
 export const typography = {
   /**
